requirements.txt
aniso8601==3.0.2
args==0.1.0
attrs==18.1.0
behave==1.2.6
certifi==2018.8.13
chardet==3.0.4
click==6.7
clint==0.5.1
coverage==4.5.1
expects==0.8.0
flask-marshmallow==0.9.0
flask-restplus==0.11.0
flask-sqlalchemy==2.3.2
flask==1.0.2
idna==2.7
inflection==0.3.1
itsdangerous==0.24
jinja2==2.10
jsonschema==2.6.0
mamba==0.9.3
markupsafe==1.0
marshmallow-sqlalchemy==0.14.1
marshmallow==2.15.4
parse-type==0.4.2
parse==1.8.4
pkginfo==1.4.2
pycodestyle==2.4.0
pyhamcrest==1.9.0
python-dateutil==2.7.3
pytz==2018.5
requests-toolbelt==0.8.0
requests==2.19.1
six==1.11.0
sqlalchemy-utils==0.33.3
sqlalchemy==1.2.10
tqdm==4.24.0
twine==1.11.0
urllib3==1.23
werkzeug==0.14.1
factory_boy==2.11.1
